Add auth controller
