+++ app/Template/layout.php
@@ -58,7 +58,7 @@
                     </li>
                     <?php endif ?>
                     <li>
-                        <?= $this->a(t('Logout'), 'user', 'logout', array(), true) ?>
+                        <?= $this->a(t('Logout'), 'auth', 'logout') ?>
                         <span class="username hide-tablet">(<?= $this->a($this->e($this->getFullname()), 'user', 'show', array('user_id' => $this->userSession->getId())) ?>)</span>
                     </li>
                 </ul>
